requirements.txt
amqp==5.3.1
annotated-types==0.7.0
ansi2html @ git+https://github.com/DMOJ/ansi2html.git@1791b4a468942eba19cb29242d9075f5e4284450
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
cbor2==5.6.5
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.3
defusedxml==0.8.0rc2
discord-webhook==1.4.1
Django==3.2.25
django-admin-sortable2==1.0.4
django-appconf==1.1.0
django-cleanup==9.0.0
django-compressor==4.4
django-fernet-fields @ git+https://github.com/DMOJ/django-fernet-fields.git@eaea7e760472106c015f65535f4837a1189c4e7c
django-impersonate==1.9.5
django-jinja==2.11.0
django-js-asset==2.2.0
django-mptt==0.17.0
django-redis==5.4.0
django-registration-redux==2.13
django-reversion==3.0.9
django-social-share==2.3.0
django-sortedm2m==4.0.0
django-statici18n==2.6.0
django-versatileimagefield==3.1
dmoj-wpadmin @ git+https://github.com/DMOJ/dmoj-wpadmin.git@dfafe1b20d69e2e0f8cf625fecbb957414040b80
et_xmlfile==2.0.0
Flask==3.1.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
icalendar==6.3.1
idna==3.10
importlib-metadata==4.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonfield @ git+https://github.com/DMOJ/jsonfield.git@560bdca857a69b4a5b2773a79d7511ebc628f339
judge==0.2.0
kombu==5.5.4
llist==0.7.1
lupa==2.4
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.8
markdown2 @ git+https://github.com/VNOI-Admin/python-markdown2.git@a5673cae919fe2eb9ea47843c009bc56b01493cb
MarkupSafe==3.0.2
martor @ git+https://github.com/CHT-OJ/martor.git@410262a812fe7884c3c1d07c2540b1666f336df8
mistune==0.8.4
mysqlclient==2.2.7
netaddr==1.3.0
numpy==2.3.2
oauthlib==3.2.2
openpyxl==3.1.3
packaging==21.3
pandas==2.3.1
pika==1.3.2
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.11.2
PyJWT==2.10.1
pymoss==0.2
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
rcssmin==1.1.1
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.7
sqlparse==0.5.3
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
uWSGI==2.0.30
vine==5.1.0
wcwidth==0.2.13
webauthn==0.4.7
webencodings==0.5.1
Werkzeug==3.1.3
zipp==3.22.0
